+++ src/automaton/core/network/http_server.cc
@@ -4,6 +4,7 @@
 #include <sstream>
 #include <string>
 #include <thread>
+
 #include <boost/asio/basic_stream_socket.hpp>
 #include <boost/asio/buffer.hpp>
 #include <boost/asio/error.hpp>
@@ -145,7 +146,7 @@ void http_session::read_header() {
 }
 
 void http_session::read_body(uint32_t body_size) {
-  uint32_t sz = body.size();
+  uint32_t sz = static_cast<uint32_t>(body.size());
   if (sz < body_size) {
     boost::asio::async_read(socket_, boost::asio::buffer(buffer, kBufferSize),
       boost::asio::transfer_exactly(body_size - sz),
